requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
pillow==11.2.1
requests==2.32.4
soupsieve==2.7
tqdm==4.67.1
typing-extensions==4.14.0
urllib3==2.5.0